+++ app/Blocks/Details/Block.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace GovUKBlogs\Blocks\Details;
+
+class Block implements \Dxw\Iguana\Registerable
+{
+	public function register()
+	{
+		add_action('init', [$this, 'registerBlock']);
+	}
+
+	public function registerBlock()
+	{
+		register_block_type(__DIR__ . '/build');
+	}
+}